refactor widget factory and dashboard
